backend/models/food_classifier.py
# import tensorflow as tf - disabled due to compatibility issues in Replit
import numpy as np
import os
from PIL import Image
import io

class FoodClassifier:
    def __init__(self):
        self.model = None
        self.labels = []
        self.model_loaded = False
        self.load_model()
        
    def load_model(self):
        """
        Load TensorFlow Lite model for food classification.
        In a production environment, this would load an actual trained model.
        For this implementation, we'll use a placeholder that simulates classification.
        """
        # In production, this would be:
        # model_path = os.path.join(os.path.dirname(__file__), '../assets/model.tflite')
        # self.interpreter = tf.lite.Interpreter(model_path=model_path)
        # self.interpreter.allocate_tensors()
        
        # Load food labels
        # For now, we'll use a small set of common foods
        self.labels = [
            "apple", "banana", "bread", "rice", "chicken", "salad", "pizza", 
            "pasta", "fish", "eggs", "milk", "cheese", "yogurt", "orange",
            "dates", "hummus", "falafel", "shawarma", "tabbouleh", "baklava"
        ]
        self.model_loaded = True
        
    def preprocess_image(self, image_data):
        """
        Preprocess the image for the model.
        """
        try:
            image = Image.open(io.BytesIO(image_data))
            image = image.resize((224, 224))
            image = image.convert('RGB')
            image_array = np.array(image)
            image_array = image_array / 255.0  # Normalize to [0,1]
            image_array = np.expand_dims(image_array, axis=0)
            return image_array
        except Exception as e:
            print(f"Error preprocessing image: {e}")
            return None
    
    def predict(self, image_data):
        """
        Predict food from image data.
        
        In a real implementation, this would use the TensorFlow Lite model.
        For now, we'll simulate predictions with a fixed set of responses.
        """
        if not self.model_loaded:
            return {"error": "Model not loaded"}
        
        # Simulate model prediction
        # In production, this would use the interpreter to run inference
        
        # Randomly select a food label and confidence score for demonstration
        import random
        random.seed(sum(image_data[:100]))  # Use image data to seed for consistency
        
        # Select up to 3 food items that might be in the image
        num_foods = random.randint(1, 3)
        selected_indices = random.sample(range(len(self.labels)), num_foods)
        
        results = []
        for idx in selected_indices:
            food = self.labels[idx]
            confidence = round(0.7 + random.random() * 0.29, 2)  # Between 0.7 and 0.99
            results.append({"food": food, "confidence": confidence})
        
        # Sort by confidence, highest first
        results.sort(key=lambda x: x["confidence"], reverse=True)
        
        return results

# Singleton instance
food_classifier = FoodClassifier()
