pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "tensorflow>=2.14.0",
    "uvicorn>=0.34.2",
]
